# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Estacionamento API
  version: 1.0.0
  description: API para registrar entradas, saídas, ocupação e histórico de estadias.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/stays/entries:
    post:
      summary: Registra a entrada de um veículo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlateRequest'
      responses:
        '201':
          description: Entrada registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingStay'
        '400':
          description: Placa inválida
        '409':
          description: Placa já possui estadia ativa ou estacionamento lotado
  /api/v1/stays/exits:
    post:
      summary: Registra a saída e calcula a cobrança
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlateRequest'
      responses:
        '200':
          description: Saída registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingStay'
        '400':
          description: Placa inválida
        '404':
          description: Não existe estadia ativa para a placa
  /api/v1/stays:
    get:
      summary: Lista o histórico de estadias
      parameters:
        - name: X-Admin-Token
          in: header
          required: false
          description: Obrigatório quando PARKING_ADMIN_TOKEN estiver configurado.
          schema:
            type: string
        - name: plate
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ACTIVE, FINISHED]
      responses:
        '200':
          description: Histórico
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParkingStay'
        '403':
          description: Token administrativo ausente ou inválido
  /api/v1/stays/active:
    get:
      summary: Lista estadias ativas
      responses:
        '200':
          description: Estadias ativas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParkingStay'
  /api/v1/stays/occupancy:
    get:
      summary: Consulta a ocupação atual
      responses:
        '200':
          description: Ocupação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Occupancy'
  /api/v1/stays/{id}:
    get:
      summary: Consulta uma estadia
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Estadia encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingStay'
        '404':
          description: Estadia não encontrada
  /api/v1/reports/operations:
    get:
      summary: Operational metrics for a date range
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
        - name: X-Admin-Token
          in: header
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Aggregated entries, exits, revenue and average stay duration
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperationalReport' }
        '403': { description: Administrative token is required }
  /api/v1/plate-detections:
    post:
      summary: Receive a provider-neutral camera or OCR plate reading
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlateDetectionRequest' }
      responses:
        '201':
          description: Reading accepted and either automatically processed or queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlateDetection' }
  /api/v1/plate-detections/pending:
    get:
      summary: List OCR readings awaiting attendant confirmation
      parameters:
        - name: X-Admin-Token
          in: header
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Pending readings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PlateDetection' }
        '403': { description: Administrative token is required }
  /api/v1/plate-detections/{id}/confirm:
    post:
      summary: Confirm a pending reading with the attendant-corrected plate
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: X-Admin-Token
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlateRequest' }
      responses:
        '200':
          description: Confirmed reading
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlateDetection' }
  /api/v1/plate-detections/{id}/dismiss:
    post:
      summary: Dismiss a pending reading without parking movement
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: X-Admin-Token
          in: header
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Dismissed reading
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlateDetection' }
components:
  schemas:
    PlateRequest:
      type: object
      required: [plate]
      properties:
        plate:
          type: string
          example: ABC1D23
    ParkingStay:
      type: object
      properties:
        id:
          type: string
          format: uuid
        plate:
          type: string
        enteredAt:
          type: string
          format: date-time
        exitedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [ACTIVE, FINISHED]
        amountCharged:
          type: number
          nullable: true
        durationMinutes:
          type: integer
          format: int64
          nullable: true
    Occupancy:
      type: object
      properties:
        capacity:
          type: integer
        occupied:
          type: integer
          format: int64
        available:
          type: integer
          format: int64
    OperationalReport:
      type: object
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        entries: { type: integer, format: int64 }
        exits: { type: integer, format: int64 }
        revenue: { type: number }
        averageStayMinutes: { type: integer, format: int64 }
    PlateDetectionRequest:
      type: object
      required: [plate, confidence, direction]
      properties:
        plate: { type: string, example: ABC1D23 }
        confidence: { type: number, minimum: 0, maximum: 1, example: 0.96 }
        direction: { type: string, enum: [ENTRY, EXIT] }
        detectedAt: { type: string, format: date-time }
    PlateDetection:
      type: object
      properties:
        id: { type: string, format: uuid }
        rawPlate: { type: string }
        confidence: { type: number }
        direction: { type: string, enum: [ENTRY, EXIT] }
        detectedAt: { type: string, format: date-time }
        status: { type: string, enum: [PENDING, PROCESSED, DISMISSED] }
        reason: { type: string, nullable: true }
        resolvedAt: { type: string, format: date-time, nullable: true }